+++ app/lib/session.js
@@ -3,10 +3,9 @@
 const session = require('express-session');
 const RedisStore = require('connect-redis')(session);
 const Redis = require('./redis');
+const linzDefaults = require('linz/lib/defaults');
 
-module.exports = session({
-    resave: false,
-    saveUninitialized: false,
-    secret: 'sessionsecret',
-    store: new RedisStore({ client: new Redis() }),
-});
+const store = new RedisStore({ client: new Redis() });
+const sessionOptions = { ...linzDefaults['session options'], ...{ store } };
+
+module.exports = session(sessionOptions);