+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## v1.0.0-18.3.0 (15 April 2019)
+
+- Added `session options` default object passed to `express-session` middleware.
+- Added `cookie options` default object passed to `cookie-parser` and `express-session` middleware.
+- The `linzReturnUrl` cookie is now signed by default.
+- Through the above, updated Linz's default cookie settings to be more inline with OWASP recommendations.
+- The example app demonstrates how to extend the default session options.
+- Removed the X-Powered-By header.
+- Added the X-XSS-Protection header.
+
 ## v1.0.0-18.2.0 (12 April 2019)
 
 - Linz's error handling middleware will now add the error to `req`.